plugin: Fixed Plugin Flashing Issues
Several errors related to the module installation and others related to the disappearance of the control center have been fixed.

Now, there is no longer a need to use the volume keys; the script automatically detects the Android version you are using.

If you have signature verification disabled, the plugin and its extension will be installed automatically as an update, and you won't need to restart the device to apply the changes.

Co-Authored-By: Kevin Miranda <83476983+KevinMiranda26@users.noreply.github.com>